minor edits, wrote code guide
as of today, the rotation is working, the speeds are as best as they can be, an i have written a guide on maintaining main.ino.

that said, main.ino has some upcoming changes!

and readme.md has some work to be done..

my dad is going to be reconfiguring the machine physically, which will allow the vertical stepper to be moving much faster. until then, i will be working on the documentation.